requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
matplotlib==3.8.3
seaborn==0.13.2
scikit-learn==1.4.2
vosk==0.3.45
sounddevice==0.4.6
scipy==1.12.0
pyttsx3==2.90
python-Levenshtein==0.23.0
openai-whisper==20231117
torch==2.2.1
transformers==4.38.2
accelerate==0.27.2
google-generativeai==0.3.2
ffmpeg-python==0.2.0 